requirements-dev.txt
###### Requirements with Version Specifiers ######


###### Requirements without Version Specifiers ######
flake8
coverage
django-cprofile-middleware
responses
